reduce complexity

+++ Logger.cs
@@ -8,87 +8,21 @@ namespace Repos.Log
     {
         // Cambie estos valores para su aplicación
         public static readonly string ReposFolder = @"c:\Repos\logs\";
-        public static readonly string LogPrefix = "repos";
+        public static string LogPrefix
+        {
+            get;
+            set;
+        } = "repos";
 
         private static Stopwatch timer;
         private static bool isStartTrivial;
         private static bool isStartDataUpdate;
-        private static bool isStartConnection; 
+        private static bool isStartConnection;
 
-        public static void Write(string mensaje)
+        private static void Writer(string file, string mensaje, ref bool firstRun, bool destroy)
         {
-            Debug.WriteLine(mensaje);
-            try
-            {
-                using (StreamWriter w = File.AppendText(ReposFolder + LogPrefix + @".log.txt"))
-                {
-                    w.WriteLine(DateTime.Now.ToString(@"yy-MM-ddtHH\:mm") + ": " + mensaje);
-                }
-            }
-            catch
-            {
-                // No se pudo escribir en el log
-            }
-        }
-        public static void WriteDataUpdate(string mensaje)
-        {
-            WriteDataUpdate(mensaje, true);
-        }
-        public static void WriteDataUpdate(string mensaje, bool hour)
-        {
-            if (!isStartDataUpdate)
-            {
-                isStartDataUpdate = true;
-                WriteDataUpdate(DateHeader, false);
-            }
-
-            Debug.WriteLine(mensaje);
-            try
-            {
-                using (StreamWriter w = File.AppendText(ReposFolder + LogPrefix + @"DataUpdate.log.txt"))
-                {
-                    string h = hour ? DateTime.Now.ToString(@"HH\:mm") + ":" : string.Empty;
-                    w.WriteLine(h + mensaje);
-                }
-            }
-            catch
-            {
-                // No se pudo escribir en el log
-            }
-        }
-        public static void WriteConnection(string mensaje)
-        {
-            WriteConnection(mensaje, true);
-        }
-        public static void WriteConnection(string mensaje, bool hour)
-        {
-            if (!isStartConnection)
-            {
-                isStartConnection = true;
-                WriteConnection(DateHeader, false);
-            }
-
-            Debug.WriteLine(mensaje);
-            try
-            {
-                using (StreamWriter w = File.AppendText(ReposFolder + LogPrefix + @"Sync.log.txt"))
-                {
-                    string h = hour ? DateTime.Now.ToString(@"HH\:mm") + ":" : string.Empty;
-                    w.WriteLine(h + mensaje);
-                }
-            }
-            catch
-            {
-                // No se pudo escribir en el log
-            }
-        }
-        public static void Trivial(string mensaje)
-        {
-            Trivial(mensaje, true);
-        }
-        public static void Trivial(string mensaje, bool hour)
-        {
-            if (!isStartTrivial)
+            string head = string.Empty;
+            if (!firstRun)
             {
                 var dir = new DirectoryInfo(ReposFolder);
                 if (!dir.Exists)
@@ -96,27 +30,24 @@ namespace Repos.Log
                     dir.Create();
                 }
 
-                timer = Stopwatch.StartNew();
-                Debug.Indent();
-                if (File.Exists(ReposFolder + LogPrefix + @"Trivial.log.txt"))
+                if (destroy)
                 {
-                    File.Delete(ReposFolder + LogPrefix + @"Trivial.log.txt");
+                    if (File.Exists(ReposFolder + LogPrefix + @"Trivial.log.txt"))
+                    {
+                        File.Delete(ReposFolder + LogPrefix + @"Trivial.log.txt");
+                    }
                 }
 
-                isStartTrivial = true;
-                Trivial(DateHeader, false);
+                firstRun = true;
+                head = "=-=-=-=- [ " + DateTime.Now.ToString(@"yy-MM-dd") + " ] =-=-=-=-" + Environment.NewLine;
             }
 
-            var mth = new StackTrace().GetFrame(1).GetMethod();
-            var _prefix = hour ? timer.Elapsed.ToString() + " (" + mth.ReflectedType.Name + ")\t\t" : string.Empty;
-            var s = _prefix + mensaje;
-
-            Debug.WriteLine(s);
+            Debug.WriteLine(mensaje);
             try
             {
-                using (StreamWriter w = File.AppendText(Logger.ReposFolder + LogPrefix + @"Trivial.log.txt"))
+                using (StreamWriter w = File.AppendText(ReposFolder + LogPrefix + file + @".log.txt"))
                 {
-                    w.WriteLine(s);
+                    w.WriteLine(head + mensaje);
                 }
             }
             catch
@@ -124,9 +55,31 @@ namespace Repos.Log
                 // No se pudo escribir en el log
             }
         }
-        private static string DateHeader => "=-=-=-=- [ " + DateTime.Now.ToString(@"yy-MM-dd") + " ] =-=-=-=-";
-        public static bool IsStartConnection { get => isStartConnection; set => isStartConnection = value; }        
-        public static bool IsStartDataUpdate { get => isStartDataUpdate; set => isStartDataUpdate = value; }
-        public static bool IsStartTrivial { get => isStartTrivial; set => isStartTrivial = value; }
+
+        public static void Write(string mensaje)
+        {
+            bool x = true;
+            var mth = new StackTrace().GetFrame(1).GetMethod().ReflectedType.Name;
+            Writer("", DateTime.Now.ToString(@"yy-MM-ddtHH\:mm") + " (" + mth + ")\t\t" + mensaje, ref x, false);
+        }
+        public static void WriteDataUpdate(string mensaje)
+        {
+            Writer("DataUpdate", DateTime.Now.ToString(@"HH\:mm") + ":" + mensaje, ref isStartDataUpdate, false);
+        }
+        public static void WriteConnection(string mensaje)
+        {
+            Writer("Sync", DateTime.Now.ToString(@"HH\:mm") + ":" + mensaje, ref isStartConnection, false);
+        }
+        public static void Trivial(string mensaje)
+        {
+            var frame = new StackTrace().GetFrame(1).GetMethod();
+            if (!isStartTrivial)
+            {
+                timer = Stopwatch.StartNew();
+                Debug.Indent();                
+            }
+            var s = timer.Elapsed.ToString() + " (" + frame + ")\t\t" + mensaje;
+            Writer("Trivial", s, ref isStartTrivial, true);
+        }
     }
 }
